+++ wifipumpkin3/__init__.py
@@ -182,13 +182,17 @@ class PumpkinShell(Qt.QObject, ConsoleUI):
         """ start access point """
         if len(self.Apthreads["RogueAP"]) > 0:
             print(display_messages("the AP is running at full power.", error=True))
-            return 
+            return
 
         self.interfaces = Linux.get_interfaces()
         if not self.conf.get(
             "accesspoint", self.commands["interface"]
         ) in self.interfaces.get("all"):
-            print(display_messages("the interface not found or is unavailable ", error=True))
+            print(
+                display_messages(
+                    "the interface not found or is unavailable ", error=True
+                )
+            )
             sys.exit(1)
 
         if self.wireless_controller.Start() != None: